requirements.txt
annotated-types==0.5.0
black==23.7.0
click==8.1.7
colorama==0.4.6
importlib-metadata==6.8.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pydantic==2.3.0
pydantic-yaml==1.1.1
pydantic_core==2.6.3
PyYAML==6.0.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
typing_extensions==4.7.1
zipp==3.16.2
